requirements.txt
flask
flask-socketio
requests
beautifulsoup4
lxml
selenium
webdriver-manager
python-dotenv